Improved blueprints design
