--- webapp/authtest.py
@@ -1,50 +1,0 @@
-### I do not think this file is till being used, check this later and remove if not needed
-
-### Was this a test file? renamed it from auth.py to authtest.py to avoid conflicts with the auth.py in the webapp folder
-
-### 
-
-
-from flask import Blueprint, render_template, redirect, url_for, request, flash
-from . import db
-from werkzeug.security import generate_password_hash, check_password_hash
-#from .models import User
-
-auth = Blueprint('auth', __name__)
-
-@auth.route('/login')
-def login():
-    return render_template('login.html')
-
-@auth.route('/signup')
-def signup():
-    return render_template('signup.html')
-
-@auth.route('/signup', methods=['POST'])
-def signup_post():
-    
-    email = request.form.get('email')
-    name = request.form.get('name')
-    password = request.form.get('password')
-
-    user = User.query.filter_by(email=email).first() # if this returns a user, then the email already exists in database
-
-    if user: # if a user is found, we want to redirect back to signup page so user can try again  
-        flash('Email address already exists')
-        return redirect(url_for('auth.signup'))
-
-    # create new user with the form data. Hash the password so plaintext version isn't saved.
-    new_user = User()
-    new_user.email = email
-    new_user.name = name
-    new_user.password = generate_password_hash(password, method='sha256')
-
-    # add the new user to the database
-    db.session.add(new_user)
-    db.session.commit()
-
-    return redirect(url_for('auth.login'))
-
-@auth.route('/logout')
-def logout():
-    return 'Logout'
