+++ webapp/settings/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('settings', __name__)
+
+from webapp.settings import routes
